Refactor SentryReporter to parse last core output for specific exception types
This commit refactors the `SentryReporter` class to parse the "last core output" only for specific exception types. The `send_event` function now checks if the exception type is in a predefined set of types that require parsing. If it is, the last core output is split for better representation in the web view and then parsed using the `parse_last_core_output` function. The resulting core exception is added to the exceptions list, along with some additional modifications to remove unnecessary information.

Additionally, a new test case has been added to ensure that the `send_event` function correctly calls `parse_last_core_output` only for specific exception types.

+++ src/tribler/core/sentry_reporter/sentry_reporter.py
@@ -93,6 +93,7 @@ class SentryReporter:
 
         self._sentry_logger_name = 'SentryReporter'
         self._logger = logging.getLogger(self._sentry_logger_name)
+        self._types_that_requires_core_output_parse = {'CoreCrashedError'}
 
     def init(self, sentry_url='', release_version='', scrubber=None,
              strategy=SentryStrategy.SEND_ALLOWED_WITH_CONFIRMATION):
@@ -205,19 +206,23 @@ class SentryReporter:
 
         # try to retrieve an error from the last_core_output
         if last_core_output:
-            # split for better representation in the web view
-            info[LAST_CORE_OUTPUT] = last_core_output.split('\n')
-            if last_core_exception := parse_last_core_output(last_core_output):
-                exceptions = event.get(EXCEPTION, {})
-                gui_exception = get_last_item(exceptions.get(VALUES, []), {})
+            info[LAST_CORE_OUTPUT] = last_core_output.split('\n')  # split for better representation in the web view
 
-                # create a core exception extracted from the last core output
-                core_exception = {TYPE: last_core_exception.type, VALUE: last_core_exception.message}
+            # check is it necessary to parse the last core output
+            exceptions = event.get(EXCEPTION, {})
+            gui_exception = get_last_item(exceptions.get(VALUES, []), {})
+            gui_exception_type = gui_exception.get(TYPE, None)
+            need_to_parse_core_output = gui_exception_type in self._types_that_requires_core_output_parse
+            if need_to_parse_core_output:
+                if last_core_exception := parse_last_core_output(last_core_output):
+                    # create a core exception extracted from the last core output
+                    core_exception = {TYPE: last_core_exception.type, VALUE: last_core_exception.message}
 
-                # remove the stacktrace field as it doesn't give any useful information for the further investigation
-                delete_item(gui_exception, 'stacktrace')
+                    # remove the stacktrace field as it doesn't give any useful information for the further
+                    # investigation
+                    delete_item(gui_exception, 'stacktrace')
 
-                exceptions[VALUES] = [gui_exception, core_exception]
+                    exceptions[VALUES] = [gui_exception, core_exception]
 
         event[CONTEXTS][REPORTER] = info
         event[CONTEXTS][BROWSER] = {VERSION: tribler_version, NAME: TRIBLER}
